Refactor ApplicationTest into route-specific files and comment existing tests
