+++ src/test/kotlin/com/flightbooking/StaffFlightsRoutesTest.kt
@@ -1,3 +1,21 @@
 package com.flightbooking
 
-class StaffFlightsRoutesTest
+import io.ktor.client.request.get
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpStatusCode
+import io.ktor.server.testing.testApplication
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class StaffFlightsRoutesTest : IntegrationTestSupport() {
+    @Test
+    // Unauthenticated staff users should be sent to the staff login page.
+    fun unauthenticatedStaffFlightsRedirectsToStaffLogin() = testApplication {
+        configureApp()
+        val client = createClient { followRedirects = false }
+
+        val response = client.get("/staff/flights")
+        assertEquals(HttpStatusCode.Found, response.status)
+        assertEquals("/staff/login", response.headers[HttpHeaders.Location])
+    }
+}
